more on issue #6

+++ lib/partition/uncoarsening/refinement/node_separators/fm_ns_local_search.cpp
@@ -52,7 +52,7 @@ EdgeWeight fm_ns_local_search::perform_refinement(const PartitionConfig & config
 
         NodeWeight best_separator  = block_weights[2];
         NodeWeight input_separator = block_weights[2];
-        int best_diff              = abs(block_weights[1]-block_weights[0]);
+        int best_diff              = abs((int)block_weights[1]-(int)block_weights[0]);
         int undo_idx = 0;
 
         int steps_till_last_improvement = 0;
@@ -103,7 +103,7 @@ EdgeWeight fm_ns_local_search::perform_refinement(const PartitionConfig & config
                         }
                 }
 
-                int cur_diff = abs(block_weights[1]-block_weights[0]);
+                int cur_diff = abs((int)block_weights[1]-(int)block_weights[0]);
                 if( block_weights[2] < best_separator || (block_weights[2] == best_separator && cur_diff < best_diff)  ) {
                         best_separator = block_weights[2];
                         undo_idx = rollback_info.size();
@@ -154,7 +154,7 @@ EdgeWeight fm_ns_local_search::perform_refinement(const PartitionConfig & config
         best_block_weights = block_weights;
         NodeWeight best_separator  = block_weights[2];
         NodeWeight input_separator = block_weights[2];
-        int best_diff              = abs(block_weights[1]-block_weights[0]);
+        int best_diff              = abs((int)block_weights[1]-(int)block_weights[0]);
         int undo_idx = 0;
 
         int steps_till_last_improvement = 0;
@@ -205,7 +205,7 @@ EdgeWeight fm_ns_local_search::perform_refinement(const PartitionConfig & config
                         }
                 }
 
-                int cur_diff = abs(block_weights[1]-block_weights[0]);
+                int cur_diff = abs((int)block_weights[1]-(int)block_weights[0]);
                 if( block_weights[2] < best_separator || (block_weights[2] == best_separator && cur_diff < best_diff)  ) {
                         best_separator              = block_weights[2];
                         undo_idx                    = rollback_info.size();
